+++ lib/ts/recipe/session/sessionFunctions.ts
@@ -30,10 +30,11 @@ export async function createNewSession(
     userId: string,
     disableAntiCsrf: boolean,
     accessTokenPayload: any = {},
-    sessionData: any = {}
+    sessionDataInDatabase: any = {}
 ): Promise<CreateOrRefreshAPIResponse> {
     accessTokenPayload = accessTokenPayload === null || accessTokenPayload === undefined ? {} : accessTokenPayload;
-    sessionData = sessionData === null || sessionData === undefined ? {} : sessionData;
+    sessionDataInDatabase =
+        sessionDataInDatabase === null || sessionDataInDatabase === undefined ? {} : sessionDataInDatabase;
 
     let requestBody: {
         userId: string;
@@ -43,7 +44,7 @@ export async function createNewSession(
     } = {
         userId,
         userDataInJWT: accessTokenPayload,
-        userDataInDatabase: sessionData,
+        userDataInDatabase: sessionDataInDatabase,
     };
 
     let handShakeInfo = await helpers.getHandshakeInfo();
@@ -286,10 +287,10 @@ export async function getSessionInformation(
 
     if (response.status === "OK") {
         // Change keys to make them more readable
-        response["sessionData"] = response.userDataInDatabase;
+        response["sessionDataInDatabase"] = response.userDataInDatabase;
         response["accessTokenPayload"] = response.userDataInJWT;
 
-        delete response.userDataInJWT;
+        delete response.userDataInDatabase;
         delete response.userDataInJWT;
 
         return response;
@@ -403,7 +404,7 @@ export async function revokeMultipleSessions(helpers: Helpers, sessionHandles: s
 /**
  * @description: It provides no locking mechanism in case other processes are updating session data for this session as well.
  */
-export async function updateSessionData(
+export async function updateSessionDataInDatabase(
     helpers: Helpers,
     sessionHandle: string,
     newSessionData: any